refactor(auth): move SqlUserRepository to sql_repository module
Keep UserRepository as the abstract contract in repository.py and extract SqlUserRepository into sql_repository.py. Update auth/repository tests imports accordingly to preserve behavior while improving extension points for additional repository implementations.
